+++ src/auth/messages.ts
@@ -1,0 +1,32 @@
+export const authMessage = {
+  chooseAuthenticationMethod: 'Authenticate in to Affinidi',
+  createAnAccountWithAffinidi: 'Create an Account with Affinidi',
+  loggedIn: 'Logged in to Affinidi',
+  loggedOut: 'Logged out of Affinidi',
+  signedUp: 'Signed up to Affinidi',
+  login: 'Log in to Affinidi',
+  loggingIn: 'Logging in to Affinidi',
+  accept: 'Accept',
+  reject: 'Reject',
+  enterEmail: 'Enter Email',
+  enterEmailOfAffinidiAccount: 'Enter the email of your Affinidi account',
+  pasteEmailAddress: 'Paste the code sent to your email',
+  invalidEmailAddress: 'Invalid email address entered',
+  emailNotFound:
+    "Please enter the email address you signed-up with or sign-up if you don't have an account",
+  confirmationCode: 'Confirmation Code',
+  sendingConfirmationCode: 'Sending confirmation code',
+  confirmationCodeSent: 'Confirmation code sent to',
+  confirmationCodeShouldBeSixDigits: 'Confirmation code should be 6 digits',
+  confirmationCodeRequired: 'Confirmation code is required',
+  createAnAccount: 'Create an Account',
+  unableToCreateSession: 'Failed to create a session',
+  noValidSessionFound: 'Valid Affinidi authentication session not found',
+  somethingWentWrong: 'Something went wrong. Please try again later.',
+  notLoggedIn: 'You are already logged out',
+  acceptedTermsAndConditions: 'You accepted terms and conditions',
+  rejectedTermsAndConditions: 'You rejected terms and conditions',
+  termsAndConditions:
+    'Please read and accept the [Terms of Use](https://build.affinidi.com/dev-tools/terms-of-use.pdf) and [Privacy Policy](https://build.affinidi.com/dev-tools/privacy-policy.pdf)',
+  fetchingAccountDetails: 'Fetching account details...',
+}
