+++ src/features/schema-manager/schema-details/showSchemaDetails.ts
@@ -4,7 +4,8 @@ import { Uri, ViewColumn, Webview, WebviewPanel, window } from 'vscode'
 import fetch from 'node-fetch'
 
 import { ext } from '../../../extensionVariables'
-import { labels, schemaMessage } from '../../../messages/messages'
+import { labels } from '../../../messages/messages'
+import { schemaMessage } from '../messages'
 import { getWebviewUri } from '../../../utils/getWebviewUri'
 import { csvCreationService } from '../../issuance/csvCreationService'
 import { schemaManagerHelpers } from '../schemaManagerHelpers'
