+++ src/features/schema-manager/schemaManagerHelpers.ts
@@ -3,7 +3,7 @@ import path from 'path'
 import { window, ProgressLocation } from 'vscode'
 import { Schema } from '../../utils/types'
 import { showQuickPick } from '../../utils/showQuickPick'
-import { schemaMessage } from '../../messages/messages'
+import { schemaMessage } from './messages'
 import { schemaManagerState } from './schemaManagerState'
 import { readOnlyContentViewer } from '../../utils/openReadOnlyContent'
 import { SchemaDto } from '@affinidi/client-schema-manager'
