+++ src/features/verify/verifyVC.ts
@@ -1,7 +1,8 @@
 import { OpenDialogOptions, ProgressLocation, window } from 'vscode'
 import fs from 'fs'
 import { ext } from '../../extensionVariables'
-import { issuanceMessage, labels } from '../../messages/messages'
+import { labels } from '../../messages/messages'
+// import { issuanceMessage } from '../messages'
 import { iamState } from '../iam/iamState'
 import { verifierClient } from './verifyClient'
 import { notifyError } from '../../utils/notifyError'
